Single-Garment-Mood-image/image-matting/image-matting-cog/requirements.txt
torch
torchvision
fairscale
timm
anyio==4.2.0
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cog==0.8.6
exceptiongroup==1.2.0
fastapi==0.98.0
gdown
h11==0.14.0
httptools==0.6.1
idna==3.6
llvmlite==0.41.1
numba==0.58.1
numpy==1.24.4
opencv-python==4.8.1.78
Pillow==10.1.0
pydantic==1.10.13
PyMatting==1.1.12
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rembg
scipy==1.10.1
sniffio==1.3.0
starlette==0.27.0
structlog==23.2.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
wandb==0.17.1